backend-session/requirements.txt
anyio==3.6.2
asgiref==3.6.0
boto3==1.28.11
botocore==1.31.11
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.51.0
django-cors-headers==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.16
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
